pyproject.toml
[tool.poetry]
name = "claimz-py"
version = "0.1.0"
description = ""
authors = ["altbdoor <lancersupraskyline@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0.3"
django-model-utils = "^4.4.0"
crispy-bootstrap4 = "^2024.1"
django-crispy-forms = "^2.1"
psycopg2-binary = "^2.9.9"
faker = "^24.14.0"

[tool.poetry.group.dev.dependencies]
djlint = "^1.34.1"
black = "^24.3.0"
isort = "^5.13.2"
poethepoet = "^0.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks.format]
shell = """
isort --profile=black .
black .
djlint claimzproject/templates/ --reformat
"""
